+++ platform/ios/platform/macos/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog for Mapbox Maps SDK for macOS
 
+## 0.17.0 - January 6, 2021
+
+### Features
+
+* Added the mbtiles file source for rendering vector tiles from file stored locally on the device.
+
 ## 0.16.0
 
 ### Styles and rendering